Create a strategic document structure (#6)
In Brazil we have a lot of kind of document, with different ways to
calculate their verification digit numbers.

The digit calculator was moved to internal folder called calculator and
it was renamed to have a name clear what kinf of calculator it is. It
will us able to have different calculators for any different document.
It was created an DigitCalculatable interface to provide a way to turn
document, such as: Cpf implementation based on Strategy Pattern.

Document struct has all behaviours about validation and how to format
the document number.

Erros is a small package with just common errors when a number is empty
or invalid.